Klondike, CardStackOfStack, Mapper 등 삭제
