+++ CardGameApp/CardGameApp/Model/CardStack/Column.swift
@@ -8,18 +8,6 @@
 
 import Foundation
 
-class Column: CardStackOfStack {
-    
-    private let position: Int
-    
-    init(position: Int) {
-        self.position = position
-    }
-    
-    override func postInfo() {
-        let userInfo = self.userInfo(position: position)
-        NotificationCenter.default.post(name: .cardStackDidChange,
-                                        object: self,
-                                        userInfo: userInfo)
-    }
+class Column: CardStack {
+
 }
