+++ CardGameApp/CardGameApp/UserInfoKey.swift
@@ -11,6 +11,4 @@ import Foundation
 struct UserInfoKey {
     
     static let cards = "cards"
-    static let stackType = "stackType"
-    static let position = "position"
 }
